data_acquisition/requirements.txt
aenum==3.1.16
affine==2.4.0
attrs==25.4.0
boto3==1.41.2
botocore==1.41.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
click-plugins==1.1.1.2
cligj==0.7.2
contourpy==1.3.3
cycler==0.12.1
cykhash==2.0.1
Cython==3.2.0
dataclasses-json==0.6.7
Deprecated==1.3.1
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
geopandas==1.1.1
idna==3.11
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
kiwisolver==1.4.9
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.7
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.5
numpy==2.3.4
oauthlib==3.3.1
openeo==0.46.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pyogrio==0.11.1
pyparsing==3.2.5
pyproj==3.7.2
pyrobuf==0.9.3
pyrosm==0.6.2
pystac==1.14.1
python-dateutil==2.9.0.post0
python-rapidjson==1.22
pytz==2025.2
rasterio==1.4.3
requests==2.32.5
requests-oauthlib==2.0.0
s3transfer==0.15.0
scikit-learn==1.7.2
scipy==1.16.3
sentinelhub==3.11.3
setuptools==80.9.0
shapely==2.1.2
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.10.16
tomli==2.3.0
tomli_w==1.2.0
torch==2.9.0
torchaudio==2.9.0
torchvision==0.24.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
utm==0.8.1
wrapt==2.0.1
xarray==2025.1.1
